scripts/livekit_agent.py
import asyncio
from livekit import rtc
from google.cloud import speech, texttospeech
from gemini_llm import GeminiLLM  # dein LLM-Modul
import os
import tempfile

class LiveKitAgent:
    def __init__(self, room_url, token):
        self.room_url = room_url
        self.token = token
        self.llm = GeminiLLM()
        self.speech_client = speech.SpeechClient()
        self.tts_client = texttospeech.TextToSpeechClient()

    async def run(self):
        async with rtc.Room.connect(self.room_url, self.token) as room:
            print("Connected to LiveKit room")
            room.on_track_subscribed = self.on_audio_track
            await asyncio.Future()  # keep running

    async def on_audio_track(self, track, publication, participant):
        print(f"Subscribed to track from {participant.identity}")

        async for audio_chunk in track:
            text = self.transcribe(audio_chunk)
            response = self.llm.complete(text)
            audio_data = self.synthesize(response)
            await self.send_audio(room=track.room, audio_data=audio_data)

    def transcribe(self, audio_chunk):
        # Beispielhafte STT-Implementierung
        audio = speech.RecognitionAudio(content=audio_chunk)
        config = speech.RecognitionConfig(
            encoding=speech.RecognitionConfig.AudioEncoding.LINEAR16,
            language_code="de-DE",
        )
        response = self.speech_client.recognize(config=config, audio=audio)
        return response.results[0].alternatives[0].transcript if response.results else ""

    def synthesize(self, text):
        input_text = texttospeech.SynthesisInput(text=text)
        voice = texttospeech.VoiceSelectionParams(
            language_code="de-DE", ssml_gender=texttospeech.SsmlVoiceGender.NEUTRAL
        )
        audio_config = texttospeech.AudioConfig(
            audio_encoding=texttospeech.AudioEncoding.LINEAR16
        )
        response = self.tts_client.synthesize_speech(
            input=input_text, voice=voice, audio_config=audio_config
        )
        return response.audio_content

    async def send_audio(self, room, audio_data):
        # Hier müsste man z. B. einen lokalen Track publizieren oder aktualisieren
        pass
